==> hw2-lasso/notebooks/1_Introduction.ipynb <==
# 1.1 Dataset Construction

import numpy as np
import pandas as pd
m = 150 #number of examples
d = 75 # dimension of features

1) Construct a random design matrix

X = np.random.rand(m*d).reshape(m,d) # Construct random design matrix X

2) Constructa true weight vector

#Initialize zeros theta with dimension d+1
theta = np.zeros(d)
#Arbituary replace the first 10 elements to 10 or -10
theta[:10] = np.random.choice([-10,10],10)

3) Construct response variable y

# Episolon: normal noice
epsilon = np.random.normal(0,0.1,m)
# Compute y
y = np.dot(X,theta)+epsilon

4) Split Training validation and testing set.

train_X = X[0:80,:]
train_y = y[0:80]
validation_X = X[80:100,:]
validation_y = y[80:100]
test_X = X[100:150,:]
test_y = y[100:150]

# Save data to local folder name p1Data
np.savetxt('../p1Data/test_X.csv',test_X,delimiter=',')
np.savetxt('../p1Data/test_y.csv',test_y,delimiter=',')
np.savetxt('../p1Data/train_X.csv',train_X,delimiter=',')
np.savetxt('../p1Data/train_y.csv',train_y,delimiter=',')
np.savetxt('../p1Data/validation_X.csv',validation_X,delimiter=',')
np.savetxt('../p1Data/validation_y.csv',validation_y,delimiter=',')
