==> hw2-lasso/notebooks/2 Ridge_Regression.ipynb <==
# Import dataset
import numpy as np
#import pandas as pd
from scipy.optimize import minimize
# train_X = pd.read_csv('../p1Data/train_X.csv',header=None)
# train_y = pd.read_csv('../p1Data/train_y.csv',header=None)
# validation_X = pd.read_csv('../p1Data/validation_X.csv',header=None)
# validation_y = pd.read_csv('../p1Data/validation_y.csv',header=None)
train_X = np.genfromtxt('../p1Data/train_X.csv',delimiter=',')
train_y = np.genfromtxt('../p1Data/train_y.csv',delimiter=',')
validation_X = np.genfromtxt('../p1Data/validation_X.csv',delimiter=',')
validation_y = np.genfromtxt('../p1Data/validation_y.csv',delimiter=',')
test_X = np.genfromtxt('../p1Data/test_X.csv',delimiter=',')
test_y = np.genfromtxt('../p1Data/test_y.csv',delimiter=',')

I use both sklearn.optimize.minimized() function and the sklearn.linear_model.Ridge to run ridge regression.

## sklear.optimize.minimize() function

from scipy.optimize import minimize

X = np.loadtxt('../data/X_train.txt')
y = np.loadtxt('../data/y_train.txt')
X_val= np.loadtxt('../data/X_valid.txt')
y_val= np.loadtxt('../data/y_valid.txt')

(N,D) = X.shape

w = np.random.rand(D,1)

def ridge(Lambda):
    def ridge_obj(theta):
        return ((np.linalg.norm(np.dot(X,theta) - y))**2)/(2*N) + Lambda*(np.linalg.norm(theta))**2
    return ridge_obj

def compute_loss(Lambda, theta):
    return ((np.linalg.norm(np.dot(X_val,theta) - y_val))**2)/(2*N)

for i in range(-5,6):
    Lambda = 10**i;
    w_opt = minimize(ridge(Lambda), w)
    print( Lambda, compute_loss(Lambda, w_opt.x))


Choose $\lambda = 1e-5$ since it achieves best valdiation square loss.

w_opt = minimize(ridge(1e-5), w)

w_opt.x

Observe that although our optimized w does gives out close-to-zero estimation to those components whose true values being zero. But in the threshold of 0.001, we do not have sparsity.  

## Compare Results with sklearn.linaer_model.Ridge

from sklearn.linear_model import Ridge

l2model = Ridge(alpha=0.1)
l2model.fit(train_X,train_y)

validation_y_predict = l2model.predict(validation_X)

#test set square loss
diff = validation_y_predict-validation_y
0.5/diff.shape[0]*np.dot(diff,diff.T)

The built in ridge function gives out similar validation set loss when alpha/lambda_reg = 0.1.
