==> hw2-lasso/notebooks/2_Ridge_Regression.ipynb <==
# Import dataset
import numpy as np
#import pandas as pd
from scipy.optimize import minimize
# train_X = pd.read_csv('../p1Data/train_X.csv',header=None)
# train_y = pd.read_csv('../p1Data/train_y.csv',header=None)
# validation_X = pd.read_csv('../p1Data/validation_X.csv',header=None)
# validation_y = pd.read_csv('../p1Data/validation_y.csv',header=None)
train_X = np.genfromtxt('../p1Data/train_X.csv',delimiter=',')
train_y = np.genfromtxt('../p1Data/train_y.csv',delimiter=',')
validation_X = np.genfromtxt('../p1Data/validation_X.csv',delimiter=',')
validation_y = np.genfromtxt('../p1Data/validation_y.csv',delimiter=',')
test_X = np.genfromtxt('../p1Data/test_X.csv',delimiter=',')
test_y = np.genfromtxt('../p1Data/test_y.csv',delimiter=',')

## sklear.optimize.minimize() function

#Define objective function of ridge regression
def l2_objective(theta,X,y,lambda_reg = 0.1):
    # X dimension: (m,d)
    # theta dimension d
    # y dimension: m
    diff = np.dot(X,theta)-y.T
    sqr_loss = 0.5/m*np.dot(diff,diff.T)
    l2_reg_term = lambda_reg*np.dot(theta,theta)
    return sqr_loss+l2_reg_term

#Function to compute the Hessian of l2 objective
def l2_hess(theta,X, y, lambda_reg=0.1):
    """
    Compute the gradient of L2-regularized square loss function given X, y and theta
    
    Args:
        X - the feature vector, 2D numpy array of size (num_instances, num_features)
        y - the label vector, 1D numpy array of size (num_instances)
        theta - the parameter vector, 1D numpy array of size (num_features)
        lambda_reg - the regularization coefficient
    
    Returns:
        grad - gradient vector, 1D numpy array of size (num_features)
    """
    diff = np.dot(X,theta.T)-y.T
    sqr_loss_grad = np.dot(diff,X)
    regularization_gradient = 2*lambda_reg*theta
    return (sqr_loss_grad+regularization_gradient)

# Dimension information
m,d = train_X.shape[0],train_X.shape[1]
# Initialize theta with ones
theta = np.ones(d)

l2_hess(theta,train_X,train_y)

minimize(fun=l2_objective,args=(train_X,train_y,0.1),method='Newton-CG',x0=theta,jac=l2_hess)

## Self-defined Ridge Regression function in Assignment 1

## Compare Results with sklearn.linaer_model.Ridge

from sklearn.linear_model import Ridge

l2model = Ridge(alpha=0.1)
l2model.fit(train_X,train_y)

test_y_predict = l2model.predict(test_X)

#test set square loss
diff = test_y_predict-test_y
1/diff.shape[0]*np.dot(diff,diff.T)
