==> hw2-lasso/notebooks/4_1.ipynb <==
# 4_1_1)

$$
\begin{split}
L'(0:v) &= \lim_{h \to 0^+} \frac{L(hv)-L(0)}{h} \\
&= \lim_{h \to 0^+} \frac{\lvert \lvert hxv-y \rvert \rvert^2+\lambda h \lvert \lvert v \rvert \rvert_1-\lvert \lvert y \rvert \rvert^2}{h}
\end{split}
$$

Apply L'Hopital Rule:
$$
\begin{split}
L'(0:v) &= (\lim_{h \to 0^+} 2(hxv - y)^Txv) +\lambda \lvert \lvert v \rvert \rvert_1\\
&= -2y^Txv + \lambda \lvert \lvert v \rvert \rvert_1
\end{split}
$$

\newpage

# 4_1_2)

If $L'(0:v) \geq 0$ we must have $\lambda \geq \frac{2y^Txv}{\lvert \lvert v \rvert \rvert_1}$. Hence the lower bound of $\lambda$ is $\frac{2y^Txv}{\lvert \lvert v \rvert \rvert_1}$

\newpage

# 4_1_3)

 The lower bound is equal to $\frac{\sum \limits_i 2 y_i x_i v_i}{\sum \limits_i \lvert v \rvert} \leq  \frac{\sum \limits_i 2 \lvert y_i x_i \rvert \lvert v_i \rvert}{\sum \limits_i \lvert v \rvert}$. This is a weighted sum of $2 \lvert y_i x_i \rvert$, which is smaller than the largest component of $2 \lvert y_i x_i \rvert$
