==> hw2-lasso/notebooks/4_2.ipynb <==
# 4_2_1)

Claim: $a$ and $b$ must have the same sign. 

Proof: Suppose $a$ and $b$ have opposite signs and are optimizer for $L$.
Let's $c = a+b$ and $d=0$
Since $x_1$ and $x_2$ are repeat features, $x_1 = x_2$,We must have $cx_1+dx_2=ax_1+bx_2$. By Cauchy equality we must have $\lambda \lvert a \rvert + \lambda \lvert b \rvert \geq \lambda \lvert a+b \rvert = \lambda \lvert c \rvert + \lambda \lvert d \rvert$. Hence c and d produce smaller L than a and b do. This is a contradiction.

Claim: If $c+d = a+b$ while $c$ and $d$ have the same sign. Then $c$ and $d$ are also optimizer for $L$.

Proof: We must have $cx_1+dx_2=ax_1+bx_2$. Since  $c$ and $d$ have the same sign, we must have $\lambda \lvert c \rvert + \lambda \lvert d \rvert = \lambda \lvert c+d \rvert = \lambda \lvert a+b \rvert = \lambda \lvert a \rvert + \lambda \lvert b \rvert$. Then c and d produce same L as a and b. Since a and b are defined to be minimizer, L must be optimized. Then c and d are also optimizer for L.

\newpage

# 4_2_2)

Claim: $a$ and $b$ must have the same sign. 

Proof: Suppose $a$ and $b$ have opposite signs and are optimizer for $L$. Let $a$ be the positive one and $b$ be the negative one.

If $a+b >0$ let $c = a+b$ and $d=0$ 

Since $x_1$ and $x_2$ are repeat features, $x_1 = x_2$,We must have $cx_1+dx_2=ax_1+bx_2$. If $a+b >0$,  Since $b$ is negative then $a+b<a$,  we have  $\lvert c \rvert =\lvert a+b \rvert \leq \lvert a \rvert$ and $\lvert d \rvert = 0 \leq \lvert b \rvert$. If $a+b <0$,  Since $a$ is positive then $a+b>b$, we have  $\lvert c \rvert =\lvert a+b \rvert \leq \lvert b \rvert$ and $\lvert d \rvert = 0 \leq \lvert a \rvert$. We can always create a new minimizer produce less L, introducing an contradiction.


Claim: $a=b$

Proof: Holding the sum $S$ of $a$ and $b$ constant. The $\lambda \lvert a \rvert + \lambda \lvert b \rvert$ is minimized when a = b. Since if both $a$ and $b$ are positive, $\lambda \lvert a \rvert + \lambda \lvert b \rvert = a^2 +b^2 = (a+b)^2 -2ab = S^2 - 2a(S-a)$ is minimized when $a = \frac{S}{2}$. The identical statement hold when both a and b are negative.
